Fix Console port handling behind a reverse proxy.
If mule is run behind a reverse proxy on a different port than on the reverse
proxy the console application will not work as it presents the mule port
instead of the reverse proxy port.

This patch uses the BIND_ALL_HOST mechanism to also set the port to the
requested / reverse proxy port.