STUDIO-12513: Git merge conflict caused by order of entries changed in mule-deploy.properties file.
Description:
Upon regenerating the mule-deploy.properties file, Studio frequently
change the order of entries of the various properties.
This causes Git conflicts when committing Mule projects to Git.
The auto generated date appears in the first line and not the first
correct self-generated comment.

Proposed Solution:
The property writing order is modified.
The writing of the date is eliminated and the header comment is added.