W-10867010: Provide a way to allow setting a default encoding
This commit adds the `apikit.defaultEncoding` system property and uses it to
replace the previous "encoding detection" mechanism provided by
`org.raml.parser.utils.StreamUtils`.

The default behaviour is still the same and it can still be specified as a
system property by using passing `guess` as the value of
`apikit.defaultEncoding`.