W-10867010: Provide a way to disable encoding detection
This commit adds the `apikit.disableEncodingGuessing` system property and uses
it to enhance the previous "encoding detection" mechanism provided by
`org.raml.parser.utils.StreamUtils`.

The default behaviour is still the same but now using
`apikit.disableEncodingGuessing=true` as a system property allows a user to use
their default charset if no encoding was specified on the Content-Type of a
request.

For more information on the default charset of a given JVM see:
https://docs.oracle.com/javase/7/docs/api/java/nio/charset/Charset.html#defaultCharset()